pyproject.toml
[tool.poetry]
name = "enacit4r-auth"
version = "0.3.0"
description = "Python authz utils for EPFL ENAC IT4R developments"
authors = ["ymarcon <yannick.marcon@epfl.ch>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-keycloak = "^4.7.0"
fastapi = "^0.115.4"
pydantic = "^2.9.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
